clang fixes.

+++ trunk/src/connection_base.cxx
@@ -913,7 +913,7 @@ string escape_param(connection_base &C,
   switch (treatment)
   {
   case treat_binary:
-    return "'" + escape_binary(string(in,len)) + "'";
+    return "'" + escape_binary(string(in, size_t(len))) + "'";
 
   case treat_string:
     return C.quote(in);
@@ -1070,9 +1070,9 @@ pqxx::result pqxx::connection_base::prepared_exec(
 #ifdef PQXX_HAVE_PQEXECPREPARED
     if (protocol_version() >= 3)
     {
-      internal::scoped_array<int> binary(nparams+1);
+      internal::scoped_array<int> binary(size_t(nparams+1));
       for (int i=0; i<expected_params; ++i)
-        binary[i] = (s.parameters[i].treatment == treat_binary);
+        binary[i] = (s.parameters[size_t(i)].treatment == treat_binary);
       for (int j=expected_params; j < nparams; ++j)
         binary[j] = (s.varargs_treatment == treat_binary);
       binary[nparams] = 0;
@@ -1102,7 +1102,8 @@ pqxx::result pqxx::connection_base::prepared_exec(
 		params[a],
 		paramlengths[a],
 		(a < expected_params) ?
-			s.parameters[a].treatment : s.varargs_treatment);
+			s.parameters[size_t(a)].treatment :
+			s.varargs_treatment);
 	if (a < nparams-1) Q << ',';
       }
       Q << ')';
@@ -1121,7 +1122,7 @@ pqxx::result pqxx::connection_base::prepared_exec(
 	           val = escape_param(*this,
 				params[n],
 				paramlengths[n],
-				s.parameters[n].treatment);
+				s.parameters[size_t(n)].treatment);
       const string::size_type keysz = key.size();
       // TODO: Skip quoted strings!  (And careful with multibyte encodings...)
       for (string::size_type h=S.find(key); h!=string::npos; h=S.find(key))
